+++ src/convenience/frameworks.node.ts
@@ -1,4 +1,5 @@
 // deno-lint-ignore-file no-explicit-any
+import { adapters as sharedAdapters } from "./frameworks.shared.ts";
 import { type IncomingMessage, type ServerResponse } from "http";
 
 const SECRET_HEADER = "X-Telegram-Bot-Api-Secret-Token";
@@ -29,45 +30,6 @@ const http = (req: IncomingMessage, res: ServerResponse) => {
     };
 };
 
-/** express web framework */
-const express = (req: any, res: any) => ({
-    update: Promise.resolve(req.body),
-    header: req.header(SECRET_HEADER),
-    end: () => res.end(),
-    respond: (json: string) => {
-        res.set("Content-Type", "application/json");
-        res.send(json);
-    },
-    unauthorized: () => {
-        res.send(401, "secret token is wrong");
-    },
-});
-
-/** koa web framework */
-const koa = (ctx: any) => ({
-    update: Promise.resolve(ctx.request.body),
-    header: ctx.get(SECRET_HEADER),
-    end: () => {
-        ctx.body = "";
-    },
-    respond: (json: string) => {
-        ctx.set("Content-Type", "application/json");
-        ctx.response.body = json;
-    },
-    unauthorized: () => {
-        ctx.status = 401;
-    },
-});
-
-/** fastify web framework */
-const fastify = (req: any, reply: any) => ({
-    update: Promise.resolve(req.body),
-    header: req.headers[SECRET_HEADER.toLowerCase()],
-    end: () => reply.status(200).send(),
-    respond: (json: string) => reply.send(json),
-    unauthorized: () => reply.code(401).send("secret token is wrong"),
-});
-
 /** worktop CloudFlare workers framework */
 const worktop = (req: any, res: any) => ({
     update: Promise.resolve(req.body.json()),
@@ -90,10 +52,8 @@ const awsLambda = (event: any, _context: any, callback: any) => ({
 export const adapters = {
     http,
     https: http,
-    express,
-    koa,
-    fastify,
     worktop,
     "aws-lambda": awsLambda,
+    ...sharedAdapters,
 };
 export const defaultAdapter = "express";
